+++ src/main.js
@@ -1,6 +1,6 @@
 import { createApp } from "vue";
 import App from "./App.vue";
-import NavBar from "./NavBar";
+import NavBar from "./components/NavBar";
 import "./registerServiceWorker";
 import router from "./router";
 import store from "./store";
